Add wrapper script around jekyll server
Due to the way Jekyll is started using two configuration files in
development, it seemed appropriate to add a wrapper script that
hides this complexity and provides a single command to start Jekyll.
This script should only be used in development, since it disables
features dedicated to production, e.g. comments.